Updates to make new local stack run

+++ apps/testclient/management/commands/create_test_user_and_application.py
@@ -26,19 +26,18 @@ def create_group(name="BlueButton"):
 
 def create_user(group, usr):
     u_name = "fred"
-    first_name = "Fred" 
+    first_name = "Fred"
     last_name = "Flinstone"
     email = "fred@example.com"
     password = "foobarfoobarfoobar"
     user_type = "BEN"
-    
+
     if usr is not None:
         u_name = usr
-        first_name = "{}{}".format(usr, "First") 
+        first_name = "{}{}".format(usr, "First")
         last_name = "{}{}".format(usr, "Last")
         email = "{}.{}@example.com".format(first_name, last_name)
         user_type = "DEV"
-
 
     if User.objects.filter(username=u_name).exists():
         User.objects.filter(username=u_name).delete()
@@ -47,17 +46,17 @@ def create_user(group, usr):
 
     if usr is not None:
         u = User.objects.create_user(username=u_name,
-                                    first_name=first_name,
-                                    last_name=last_name,
-                                    email=email)
+                                     first_name=first_name,
+                                     last_name=last_name,
+                                     email=email)
         u.set_unusable_password()
     else:
         # create a sample user 'fred' for dev local that has a usable password
         u = User.objects.create_user(username=u_name,
-                                    first_name=first_name,
-                                    last_name=last_name,
-                                    email=email,
-                                    password=password,)
+                                     first_name=first_name,
+                                     last_name=last_name,
+                                     email=email,
+                                     password=password,)
 
     UserProfile.objects.create(user=u,
                                user_type=user_type,
@@ -72,29 +71,60 @@ def create_user(group, usr):
     u.groups.add(group)
 
     if usr is None:
-        c, g_o_c = Crosswalk.objects.get_or_create(user=u,
-                                                   fhir_id_v2=settings.DEFAULT_SAMPLE_FHIR_ID_V2,
-                                                   _user_id_hash="ee78989d1d9ba0b98f3cfbd52479f10c7631679c17563186f70fbef038cc9536")
+        if not Crosswalk.objects.filter(_user_id_hash="ee78989d1d9ba0b98f3cfbd52479f10c7631679c17563186f70fbef038cc9536").exists():
+            c, g_o_c = Crosswalk.objects.get_or_create(user=u,
+                                                       fhir_id_v2=settings.DEFAULT_SAMPLE_FHIR_ID_V2,
+                                                       _user_id_hash="ee78989d1d9ba0b98f3cfbd52479f10c7631679c17563186f70fbef038cc9536")
+        else:
+            print("Skipping crosswalk creation; already exists.")
     return u
 
 
-def create_application(user, group, app, redirect):
-    app_name = "TestApp" if app is None else app
+# def create_application(user, group, app, redirect):
+#     app_name = "TestApp" if app is None else app
+#     Application.objects.filter(name=app_name).delete()
+#     redirect_uri = "{}{}".format(settings.HOSTNAME_URL, settings.TESTCLIENT_REDIRECT_URI)
+
+#     if redirect:
+#         redirect_uri = redirect
+
+#     if not (redirect_uri.startswith("http://") or redirect_uri.startswith("https://")):
+#         redirect_uri = "https://" + redirect_uri
+
+#     a = Application.objects.create(name=app_name,
+#                                    redirect_uris=redirect_uri,
+#                                    user=user,
+#                                    data_access_type="THIRTEEN_MONTH",
+#                                    client_type="confidential",
+#                                    authorization_grant_type="authorization-code")
+
+#     titles = ["My Medicare and supplemental coverage information.",
+#               "My Medicare claim information.",
+#               "My general patient and demographic information.",
+#               "Profile information including name and email."
+#               ]
+
+#     for t in titles:
+#         c = ProtectedCapability.objects.get(title=t)
+#         a.scope.add(c)
+#     return a
+
+def create_application(user):
+    app_name = "TestApp"
+    if Application.objects.filter(name=app_name).exists():
+        return Application.objects.get(name=app_name)
+
+    # If the app doesn't exist, create the test app.
+
     Application.objects.filter(name=app_name).delete()
     redirect_uri = "{}{}".format(settings.HOSTNAME_URL, settings.TESTCLIENT_REDIRECT_URI)
 
-    if redirect:
-        redirect_uri = redirect
-
-    if not(redirect_uri.startswith("http://") or redirect_uri.startswith("https://")):
-        redirect_uri = "https://" + redirect_uri
-
-    a = Application.objects.create(name=app_name,
-                                redirect_uris=redirect_uri,
-                                user=user,
-                                data_access_type="THIRTEEN_MONTH",
-                                client_type="confidential",
-                                authorization_grant_type="authorization-code")
+    the_app = Application.objects.create(name=app_name,
+                                         redirect_uris=redirect_uri,
+                                         user=user,
+                                         data_access_type="THIRTEEN_MONTH",
+                                         client_type="confidential",
+                                         authorization_grant_type="authorization-code",)
 
     titles = ["My Medicare and supplemental coverage information.",
               "My Medicare claim information.",
@@ -104,8 +134,9 @@ def create_application(user, group, app, redirect):
 
     for t in titles:
         c = ProtectedCapability.objects.get(title=t)
-        a.scope.add(c)
-    return a
+        the_app.scope.add(c)
+
+    return the_app
 
 
 def create_test_token(user, application):
@@ -118,10 +149,13 @@ def create_test_token(user, application):
     for s in scopes:
         scope.append(s.slug)
 
-    t = AccessToken.objects.create(user=user, application=application,
-                                   token="sample-token-string",
-                                   expires=expires,
-                                   scope=' '.join(scope))
+    if AccessToken.objects.filter(token="sample-token-string").exists():
+        t = AccessToken.objects.get(token="sample-token-string")
+    else:
+        t = AccessToken.objects.create(user=user, application=application,
+                                       token="sample-token-string",
+                                       expires=expires,
+                                       scope=' '.join(scope))
     return t
 
 
@@ -140,7 +174,7 @@ class Command(BaseCommand):
 
         g = create_group()
         u = create_user(g, usr)
-        a = create_application(u, g, app, redirect)
+        a = create_application(u)  # a = create_application(u, g, app, redirect)
         t = None
         if usr is None and app is None:
             t = create_test_token(u, a)
